Assignment_Week_8/requirements.txt
streamlit
pandas
sentence-transformers
faiss-cpu
openai
transformers
